requirements.txt
python-telegram-bot==13.15
SQLAlchemy>=2.0,<3.0
Flask>=2.0.0
requests>=2.25.0
pytz>=2021.1
